Day 6 Part 2

+++ day6/src/Race.java
@@ -1,17 +1,17 @@
 
 public class Race {
-    private final int time;
-    private final int record;
+    private final long time;
+    private final long record;
     private int nWaysToBeat;
 
-    public Race(int time, int record) {
+    public Race(long time, long record) {
         this.time = time;
         this.record = record;
         this.nWaysToBeat = 0;
     }
 
     public int calculateNWaysToBeatRecord() {
-        for (int i = 1; i <= time; i++) {
+        for (long i = 1; i <= time; i++) {
             var timeAfterReleasing = time - i;
             var distanceInTry = i * timeAfterReleasing;
             if (distanceInTry > record) {
